api/requirements.txt
asgiref==3.11.1
cryptography>=43.0.0
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1

coverage==7.8.0
Pillow>=11.0.0
pyotp==2.9.0
qrcode[pil]==8.0
sqlparse==0.5.5
gunicorn==23.0.0
psycopg2==2.9.10
whitenoise==6.11.0
celery[redis]>=5.4.0
redis>=5.0.0
Jinja2>=3.1.0
pyzipper>=0.3.6
concurrent-log-handler==0.9.25
